office_tool/backend/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.1
msal==1.31.0
requests==2.32.3
PyPDF2==3.0.1
reportlab==4.0.7
xhtml2pdf==0.2.13

SQLAlchemy==2.0.36
psycopg[binary]

gunicorn==21.2.0
google-auth
google-auth-oauthlib
google-api-python-client
